move most of auto-docs code into own project
