+++ libs/auto-docs/src/init.ts
