+++ libs/auto-docs/src/routes.ts
