+++ references/routes.json
@@ -56,31 +56,7 @@
     "method": "get"
   },
   {
-    "route": "/module-condensed/:moduleCode",
-    "method": "get"
-  },
-  {
     "route": "/module-full/:moduleCode",
-    "method": "get"
-  },
-  {
-    "route": "/module/:moduleCode",
-    "method": "get"
-  },
-  {
-    "route": "/modules",
-    "method": "get"
-  },
-  {
-    "route": "/modules-condensed",
-    "method": "get"
-  },
-  {
-    "route": "/search/modules-condensed/:searchQuery",
-    "method": "get"
-  },
-  {
-    "route": "/search/modules/:searchQuery",
     "method": "get"
   },
   {
